+++ sdk/ml/azure-ai-ml/azure/ai/ml/_schema/pipeline/pipeline_parallel_job.py
@@ -4,13 +4,12 @@
 
 import logging
 from typing import Any
-from marshmallow import fields, post_load, INCLUDE
+from marshmallow import fields, post_load
 from azure.ai.ml._schema.core.fields import NestedField, UnionField
 from azure.ai.ml._schema.job.input_output_entry import OutputSchema
 
 from azure.ai.ml._schema.job.parallel_job import ParallelJobSchema
-from azure.ai.ml._schema._deployment.batch.batch_deployment_settings import BatchRetrySettingsSchema
-from azure.ai.ml._schema.core.fields import ComputeField, ArmVersionedStr
+from azure.ai.ml._schema.core.fields import ComputeField, ArmVersionedStr, RegistryStr
 from azure.ai.ml._schema.assets.environment import AnonymousEnvironmentSchema
 from azure.ai.ml.constants import AzureMLResourceType
 
@@ -21,6 +20,7 @@ class PipelineParallelJobSchema(ParallelJobSchema):
     compute = ComputeField()
     environment = UnionField(
         [
+            RegistryStr(azureml_type=AzureMLResourceType.ENVIRONMENT),
             NestedField(AnonymousEnvironmentSchema),
             ArmVersionedStr(azureml_type=AzureMLResourceType.ENVIRONMENT, allow_default_version=True),
         ],